+++ lib/queries/serial/autocomplete.rb
@@ -1,5 +1,5 @@
 module Queries
-  module Serial 
+  module Serial
     class Autocomplete < Queries::Query
 
       # @param [Hash] args
@@ -11,7 +11,7 @@ module Queries
       def base_query
         ::Serial.select('serials.*')
       end
- 
+
       # @return [Array]
       def autocomplete
         queries = [
@@ -38,12 +38,12 @@ module Queries
           a = q
           if project_id
             # a = q.joins(:project_sources).where(member_of_project_id.to_sql) if project_id && limit_to_project
-            a = a.select('serials.*, COUNT(project_sources.source_id) AS use_count, NULLIF(project_sources.project_id, NULL) as in_project')
+            a = a.select("serials.*, COUNT(project_sources.source_id) AS use_count, CASE WHEN project_sources.project_id = #{Current.project_id} THEN project_sources.project_id ELSE NULL END AS in_project")
                  .left_outer_joins(:sources)
                  .joins('LEFT OUTER JOIN project_sources ON sources.id = project_sources.source_id')
                  .where('project_sources.project_id = ? OR project_sources.project_id IS DISTINCT FROM ?', project_id, project_id)
                  .group('serials.id, project_sources.project_id')
-                 .order('use_count DESC')
+                 .order('in_project, use_count DESC')
           end
           a ||= q
           updated_queries[i] = a
@@ -63,16 +63,16 @@ module Queries
       end
 
       # @return [ActiveRecord::Relation]
-      #    match exact name 
+      #    match exact name
       def autocomplete_exact_name
-        a = table[:name].eq(query_string) 
+        a = table[:name].eq(query_string)
         base_query.where(a.to_sql).limit(20)
       end
 
       # @return [ActiveRecord::Relation]
       #    match begining of name
       def autocomplete_begining_name
-        a = table[:name].matches(query_string + '%') 
+        a = table[:name].matches(query_string + '%')
         base_query.where(a.to_sql).limit(20)
       end
 
@@ -113,7 +113,7 @@ module Queries
       #      .on(table[:id].eq(alternate_value_table[:alternate_value_object_id]).and(
       #        alternate_value_table[:alternate_value_object_type].eq('Serial')
       #      )
-      #    ) 
+      #    )
       #  end
 
       # @return [Arel::Table]